requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-dotenv